+++ src/NxGrid/NxGrid.Sorting.cs
@@ -82,7 +82,7 @@ public partial class NxGrid<T>
 
     private async Task OnColumnClick(NxGridColumn<T> column)
     {
-        if (HeaderClickSelects) return;
+        if (HeaderClickSelects || !column.CanSort) return;
         if (isEditing) await CommitEdit();
 
         if (IsPrimarySort(column))